build: Fix all header files in the project
* Clean-up inclusions
* Apply an uniform code style
* Prepared public headers to be shipped with the library
  * Create a header for base classes of each component
  * Move base classes out of common
  * Move Pose from math to common
  * Make math.hh internal only
  * Create an auto-generated global header
    to include all other public headers
