+++ src/avoidance/Avoidance.hh
@@ -1,0 +1,32 @@
+/*
+ * Copyright (c) 2017 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include <memory>
+#include <vector>
+
+#include "common/common.hh"
+
+template <typename VehicleType>
+class CollisionAvoidanceStrategy
+{
+public:
+    virtual void avoid(const std::vector<Obstacle> &elements) = 0;
+
+protected:
+    std::shared_ptr<VehicleType> vehicle;
+};
