+++ src/avoidance/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
 
 set(SOURCES
     QuadCopterShiftAvoidance.cc
-    QuadCopterVFFAvoidance.cc
-    QuadCopterStopAvoidance.cc)
+    QuadCopterStopAvoidance.cc
+    QuadCopterVFFAvoidance.cc)
 
 add_library(avoidance OBJECT ${SOURCES})
