+++ src/common/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
 
 set(SOURCES
-  math.cc
-  DepthCamera.cc)
+    common.cc
+    math.cc)
 
 add_library(common OBJECT ${SOURCES})
