+++ src/detection/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.1 FATAL_ERROR)
 
 set(SOURCES
     DepthImageObstacleDetector.cc
-    DepthImageStraightLineDetector.cc
-    DepthImagePolarHistDetector.cc)
+    DepthImagePolarHistDetector.cc
+    DepthImageStraightLineDetector.cc)
 
 add_library(detection OBJECT ${SOURCES})
