+++ src/sensors/GazeboRealSenseCamera.cc
@@ -1,4 +1,20 @@
-#include "GazeboRealSenseCamera.hh"
+/*
+ * Copyright (c) 2017 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "sensors/GazeboRealSenseCamera.hh"
 
 // TODO: The camera configuration should be aggregated on the gazebo .sdf model
 // file or retrieved from the librealsense. This feature is not implemented
@@ -56,4 +72,3 @@ void GazeboRealSenseCamera::on_stream_depth_recvd(ConstImageStampedPtr &_msg)
     uint buffer_size = _msg->image().width() * _msg->image().height();
     depth_buffer[this->current_buffer ^ 1] = std::vector<uint16_t>(data, data + buffer_size);
 }
-
