+++ src/sensors/Sensors.hh
@@ -13,6 +13,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 */
+
 #pragma once
 
 #include <cstdint>
@@ -20,18 +21,19 @@
 
 class DepthCamera
 {
-    protected:
-        unsigned int width = 0;
-        unsigned int height = 0;
-        double scale = 0;
-        double hfov = 0.0;
-        double vfov = 0.0;
+public:
+    unsigned int get_height();
+    unsigned int get_width();
+    double get_scale();
 
-    public:
-        virtual std::vector<uint16_t> &get_depth_buffer() = 0;
-        unsigned int get_height();
-        unsigned int get_width();
-        double get_scale();
-        virtual double get_horizontal_fov() { return hfov; };
-        virtual double get_vertical_fov() { return vfov; };
+    virtual std::vector<uint16_t> &get_depth_buffer() = 0;
+    virtual double get_horizontal_fov() { return hfov; };
+    virtual double get_vertical_fov() { return vfov; };
+
+protected:
+    unsigned int height = 0;
+    unsigned int width = 0;
+    double scale = 0;
+    double hfov = 0.0;
+    double vfov = 0.0;
 };
