+++ src/vehicles/Vehicles.hh
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2017 Intel Corporation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include "common/common.hh"
+
+class QuadCopter
+{
+public:
+    virtual Pose target_pose() = 0;
+    virtual Pose vehicle_pose() = 0;
+    virtual void set_target_pose(Pose pose) = 0;
+};
